fix: ensure validation is run on correct files
